Get clojurescript working (WIP)
Undo change to declare logging functions in a loop.

That approach fails in clojurescript because eval generally isn't
available.

Can't :import the reference to the Logger protocol. Needed to :require
and :refer it instead.

Worked on the unit tests, but still having issues with the reference
to log! which is part of that protocol.

I'm not sure what those "issues" entail. The stack trace is ~1900
lines of NO_SOURCE_FILE, which reminded me that I never got around to
sorting out how mapped-stacktrace-str is precisely supposed to work.

And also: renamed `log-atomically` to
`log-and-flush-atomically!`. Broke the isolated log accumulation part
of that out into `log-atomically!`

Updated the README and CHANGELOG to reflect this.